res/requirements.txt
Flask
kafka-python
simplejson